requirements.txt
pandas>=2.0.0
numpy>=1.24.0
google-cloud-bigquery>=3.10.0
openai>=1.0.0
pyyaml>=6.0
requests>=2.31.0
jupyter>=1.0.0
ipython>=8.0.0
